refactor: move internals of commit.py to commit_utils.py
This moves:
- CommitScannable
- _parse_patch() (becomes parse_patch())
- Related constants
